finished visualize_fitting_result
