+++ menpofit/visualize/__init__.py
@@ -1,3 +1,3 @@
 from .widgets import (visualize_shape_model, visualize_appearance_model,
-                      visualize_aam, visualize_atm, plot_ced)
-#                      visualize_fitting_results,
+                      visualize_aam, visualize_atm, plot_ced,
+                      visualize_fitting_results)
