+++ menpofit/visualize/widgets/__init__.py
@@ -1,5 +1,5 @@
 from .base import (visualize_shape_model, visualize_appearance_model,
-                   visualize_aam, visualize_atm, plot_ced)
-#                   visualize_fitting_results,
+                   visualize_aam, visualize_atm, plot_ced,
+                   visualize_fitting_results)
 from .options import (LinearModelParametersWidget, FittingResultOptionsWidget,
                       FittingResultIterationsOptionsWidget)
